Fix several cyclical type imports
Move us closer to enabling the `import/no-cycle` rule by fixing several
violations.
