+++ src/components/DataTable/components/Cell/Cell.tsx
@@ -4,7 +4,7 @@ import {classNames} from '@shopify/react-utilities/styles';
 import {headerCell} from '../../../shared';
 import {withAppProvider, WithAppProviderProps} from '../../../AppProvider';
 import Icon, {IconSource} from '../../../Icon';
-import {SortDirection} from '../../DataTable';
+import {SortDirection} from '../../types';
 
 import * as styles from '../../DataTable.scss';
