+++ src/components/DataTable/index.ts
@@ -1,11 +1,7 @@
 import DataTable from './DataTable';
 
-export {
-  Props,
-  TableRow,
-  TableData,
-  SortDirection,
-  ColumnContentType,
-} from './DataTable';
+export {SortDirection} from './types';
+
+export {Props, TableRow, TableData, ColumnContentType} from './DataTable';
 
 export default DataTable;
