+++ src/components/ResourceList/types.ts
@@ -1,3 +1,15 @@
 export type SelectedItems = string[] | 'All';
 
 export const SELECT_ALL_ITEMS = 'All';
+
+export interface ResourceListContext {
+  selectMode: boolean;
+  selectable?: boolean;
+  selectedItems?: SelectedItems;
+  resourceName: {
+    singular: string;
+    plural: string;
+  };
+  loading?: boolean;
+  onSelectionChange?(selected: boolean, id: string): void;
+}
