+++ src/components/Tabs/utilities.ts
@@ -1,4 +1,4 @@
-import {TabDescriptor} from './Tabs';
+import {TabDescriptor} from './types';
 
 export function getVisibleAndHiddenTabIndices(
   tabs: TabDescriptor[],
